src/requirements.txt
# Core dependencies for Spirit Airlines BI Agent
chainlit>=2.7.0,<3.0.0
python-dotenv>=1.0.0
databricks-sdk>=0.65.0

# Database drivers (PostgreSQL only)
psycopg2-binary>=2.9.0
psycopg[binary]==3.2.9
asyncpg>=0.29.0
sqlalchemy==2.0.41
greenlet>=3.0.0

# Databricks authentication dependencies
streamlit>=1.28.0
openai>=1.0.0

# Additional utilities
python-dateutil>=2.8.0
asyncio-mqtt>=0.16.0

# Production dependencies
gunicorn>=21.0.0
uvicorn>=0.24.0
